// src/app/shared/generated/model/createSkillRequest.ts
/**
 * onecx-ai-provider  service
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface CreateSkillRequest { 
    name?: string;
    description?: string;
    instruction?: string;
}